+++ variable_set.go
@@ -72,7 +72,7 @@ type VariableSet struct {
 	Description string `jsonapi:"attr,description"`
 	Global      bool   `jsonapi:"attr,global"`
 	// **Note: This field is still in BETA and subject to change.**
-	Enforced bool `jsonapi:"attr,enforced"`
+	Priority bool `jsonapi:"attr,priority"`
 
 	// Relations
 	Organization *Organization          `jsonapi:"relation,organization"`
@@ -117,8 +117,9 @@ type VariableSetCreateOptions struct {
 	Global *bool `jsonapi:"attr,global,omitempty"`
 
 	// **Note: This field is still in BETA and subject to change.**
-	// If true the variables in the set cannot be overwritten and take precedence.
-	Enforced *bool `jsonapi:"attr,enforced,omitempty"`
+	// If true the variables in the set override any other variable values set
+	// in a more specific scope including values set on the command line.
+	Priority *bool `jsonapi:"attr,priority,omitempty"`
 }
 
 // VariableSetReadOptions represents the options for reading variable sets.
@@ -146,8 +147,9 @@ type VariableSetUpdateOptions struct {
 	Global *bool `jsonapi:"attr,global,omitempty"`
 
 	// **Note: This field is still in BETA and subject to change.**
-	// If true the variables in the set cannot be overwritten and take precedence.
-	Enforced *bool `jsonapi:"attr,enforced,omitempty"`
+	// If true the variables in the set override any other variable values set
+	// in a more specific scope including values set on the command line.
+	Priority *bool `jsonapi:"attr,priority,omitempty"`
 }
 
 // VariableSetApplyToWorkspacesOptions represents the options for applying variable sets to workspaces.